+++ RegistroArticulo/Entidades/Cotizaciones.cs
@@ -12,11 +12,11 @@ namespace RegistroArticulo.Entidades
         public int ID { get; set; }
         public DateTime Fecha { get; set; }
         public string Comentario { get; set; }
-        public decimal Monto { get; set; }
+        public int Monto { get; set; }
 
         public virtual List<CotizacionesDetalle> CotizacionDetalle { get; set; }
 
-        public Cotizaciones(DateTime fecha, string comentario, decimal monto, List<CotizacionesDetalle> Cotizaciondetalle)
+        public Cotizaciones(DateTime fecha, string comentario, int monto, List<CotizacionesDetalle> Cotizaciondetalle)
         {
             this.Fecha = fecha;
             this.Comentario = comentario;
